% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{addSampleState}
\alias{addSampleState}
\title{Add WGD state to data frame}
\usage{
addSampleState(results_MajorAP, results_MinorAP)
}
\arguments{
\item{results_MajorAP}{Data frame with major allele ploidy data for each autosome arm of each sample and annotation information}

\item{results_MinorAP}{Data frame with minor allele ploidy data for each autosome arm of each sample and annotation information}
}
\value{
results_MajorAP
}
\description{
Determines the Whole Genome Duplication (WGD) status for each sample and adds it to data_frame
}
\examples{
addSampleState(results_MajorAP, results_MinorAP)
}
