% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{addTypeOfCancer}
\alias{addTypeOfCancer}
\title{Add primary tumour location}
\usage{
addTypeOfCancer(data_frame, annoCodes)
}
\arguments{
\item{data_frame}{Data frame which has sampleID's as rownames}

\item{annoCodes}{Data frame which contains sampleID's and their primary tumour location}
}
\value{
tmpFrame
}
\description{
Merges data frames by row names. Particularly annoCodes and results_Major/Minor or CopyNumber data frame
}
\examples{
addTypeOfCancer(results_CopyNumber, annoCodes)
addTypeOfCancer(results_MajorAP,    annoCodes)
addTypeOfCancer(results_MinorAP,    annoCodes)
}
