% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{aneuploidyScores}
\alias{aneuploidyScores}
\title{Get gains, losses and aneuploidyScore (gains + losses) for each sample}
\usage{
aneuploidyScores(results_CopyNumber, valNorm)
}
\arguments{
\item{results_CopyNumber}{Data frame with copy number data for each autosome arm of each sample and annotation information}

\item{valNorm}{Can be either "Genome" or "WGD" and indicates how to normalize the data}
}
\value{
aneuploidyScoreDF
}
\description{
Determines the number of arms which had a gain or a loss and adds data to data frame aneuploidyScoreDF
}
\examples{
aneuploidyScores(results_CopyNumber, "Genome")
aneuploidyScores(results_CopyNumber, "WGD")
}
