% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_boxPlots.R
\name{cancerTypes}
\alias{cancerTypes}
\title{Whole genome duplication distribution per cancer type barplot}
\usage{
cancerTypes(results_CopyNumber)
}
\arguments{
\item{results_CopyNumber}{Data frame which contains copy number data for each autosome per sample}
}
\value{
None
}
\description{
For each cancer type the contribution of whole genome duplication statuses is determined and visualized
}
\examples{
cancerTypes(results_CopyNumber)
}
