% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{checkDataAndRepair}
\alias{checkDataAndRepair}
\title{Data check and repair}
\usage{
checkDataAndRepair(data_frame)
}
\arguments{
\item{data_frame}{Can be either one of the results_ data frames}
}
\value{
data_frame Inserted data frame but without the NA values
}
\description{
Checks data for NA's in genome_level column and replaces them with most common value of autosomes
Identifies positions in df with wrong primaryTumorLocation and changes them to "unknown"
}
\examples{
checkDataAndRepair(results_CopyNumber)
checkDataAndRepair(results_MajorAP)
checkDataAndRepair(results_MinorAP)
}
