% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{cleanData}
\alias{cleanData}
\title{Clean data}
\usage{
cleanData(data_frame)
}
\arguments{
\item{data_frame}{Can be either one of results_ data frames}
}
\value{
data_frame
}
\description{
Filters out all cancertypes with a prevalence less or equal than 10 and the unknown or wrongly classified
}
\examples{
cleanData(results_CopyNumber)
cleanData(results_MajorAP)
cleanData(results_MinorAP)
}
