% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_boxPlots.R
\name{duplicated_arms_plot}
\alias{duplicated_arms_plot}
\title{Duplicated autosome arms plot}
\usage{
duplicated_arms_plot(results_MajorAP)
}
\arguments{
\item{results_MajorAP}{Data frame with major allele ploidy data for each autosome arm of each sample and annotation information}
}
\value{
None
}
\description{
Plots number of autosome arms duplicated on xaxis and the number of samples with the same number on yaxis.
}
\examples{
duplicated_arms_plot(results_MajorAP)
}
