% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{getAneuploidyDF}
\alias{getAneuploidyDF}
\title{Add mutational load data to aneuploidyScoreDF}
\usage{
getAneuploidyDF(aneuploidyScoreDF, mut_load_df, clonal_mut_load_df)
}
\arguments{
\item{aneuploidyScoreDF}{Data frame which contains gains, losses and aneuploidy score (gains + losses) per sample}

\item{mut_load_df}{Data frame which contains mutational load data per sample}

\item{clonal_mut_load_df}{Data frame which contains clonal and subclonal mutational load data per sample}
}
\value{
aneuploidyScoreDF
}
\description{
Adds mutational load data to aneuploidyScoreDF
}
\examples{
getAneuploidyDF(aneuploidyScoreDF, mut_load_df, clonal_mut_load_df)
}
