% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{getChromArms}
\alias{getChromArms}
\title{Character string with chromosome arms and genome_level}
\usage{
getChromArms()
}
\value{
chromArms A character vector with autosome arms
}
\description{
Creates character vector with autosome arms
}
\examples{
getChromArms()
}
