% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{getCount}
\alias{getCount}
\title{Get loss or gain count}
\usage{
getCount(data_frame, type)
}
\arguments{
\item{data_frame}{Data frame with numercial data}

\item{type}{Defines if gains or losses need to be counted.}
}
\value{
return_list List with the total losses or gains per data frame column
}
\description{
Counts the total losses or gains for the data frame depending on given type (loss or gain)
}
\examples{
getCount(results_CopyNumber, "gain")
getCount(results_CopyNumber, "loss")
getCount(results_MajorAP,    "gain")
getCount(results_MajorAP,    "loss")
getCount(results_MinorAP,    "gain")
getCount(results_MinorAP,    "loss")
}
