% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{getMeans}
\alias{getMeans}
\title{Get mean of columns}
\usage{
getMeans(data_frame)
}
\arguments{
\item{data_frame}{Can be either one of the results_ data frames}
}
\value{
mean_list List with mean values of the columns of data frame
}
\description{
Calculates mean of every row in the data frame that is given as parameter
}
\examples{
getMeans(results_CopyNumber)
getMeans(results_MajorAP)
getMeans(results_MinorAP)
}
