% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{getMutloadAnnotations}
\alias{getMutloadAnnotations}
\title{Get mutational load annotations}
\usage{
getMutloadAnnotations(mut_load_df, clonal_mut_load_df,
  sign_contribution_df)
}
\arguments{
\item{mut_load_df}{Data frame which contains mutational load data per sample}

\item{clonal_mut_load_df}{Data frame which contains clonal and subclonal mutational load data per sample}

\item{sign_contribution_df}{Data frame which contains the signature contribution for each signature per sample}
}
\value{
mut_load_anno Data frame with mutational load annotation per sample
}
\description{
Constructs one mutational load data frame for annotations per sample of three single mutational load data frames
}
\examples{
getMutloadAnnotations(mut_load_df, clonal_mut_load_df, sign_contribution_df)
}
