% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{getPercentage}
\alias{getPercentage}
\title{Get percentage of 0 values in column}
\usage{
getPercentage(data_frame)
}
\arguments{
\item{data_frame}{Can be any given numerical data frame}
}
\value{
tmp_list List with percentages of each column in data frame
}
\description{
Calculates the percentage of 0 in each column of given data frame and returns
list with all percentages of each column in given data frame
}
\examples{
getPercentage(results_CopyNumber)
getPercentage(results_MajorAP)
getPercentage(results_MinorAP)
}
