% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainPloidyAnalyzer.R
\name{hclust2}
\alias{hclust2}
\title{Clustering function}
\usage{
hclust2(data_frame)
}
\arguments{
\item{data_frame}{The data frame which needs to be clustered}
}
\value{
None
}
\description{
Average cluster function for clustering the unsupervised heatmaps
}
\examples{
hclust()
}
