% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapPlots.R
\name{heatMap_plot}
\alias{heatMap_plot}
\title{Plot function heatmap}
\usage{
heatMap_plot(heatmap_df, plotName)
}
\arguments{
\item{heatmap_df}{Data frame which contains the data for heatmap plot.}

\item{plotName}{character string which is used as plot name and determines the the number of breaks for the heatmap}
}
\value{
None
}
\description{
Makes heatmap plot of data with and without a dendogram clustering
}
\examples{
heatMap_plot(heatmap_df, "CopyNumber not WGD")
heatMap_plot(heatmap_df, "CopyNumber WGD")
heatMap_plot(heatmap_df, "CopyNumber")
heatMap_plot(heatmap_df, "loh")
heatMap_plot(heatmap_df, "major")
heatMap_plot(heatmap_df, "minor")
heatMap_plot(heatmap_df, "major/minor")
}
