% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapPlots.R
\name{heatPlotsUnsupervised}
\alias{heatPlotsUnsupervised}
\title{Plot function unsupervised heatmap}
\usage{
heatPlotsUnsupervised(data_frame, annoColors, annoDF, typeColors, tp)
}
\arguments{
\item{data_frame}{Data frame that contains numerical values to plot heatmap}

\item{annoColors}{Data frame which contains per type of value the right colour code}

\item{annoDF}{Data frame that contains annotation information}

\item{typeColors}{Character string that contains 21 color codes (for each cancertype one)}

\item{tp}{Character string that defines the plot name}
}
\value{
None
}
\description{
Plot function for unsupervised heatmap of copy number data
}
\examples{
heatPlotUnsupervised(data_frame, annoColors, annoDF, typeColors, "All")
heatPlotUnsupervised(data_frame, annoColors, annoDF, typeColors, "Non WGD")
heatPlotUnsupervised(data_frame, annoColors, annoDF, typeColors, "WGD")
}
