% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_boxPlots.R
\name{lohDistribution}
\alias{lohDistribution}
\title{Loss of Heterozygosity distribution barplot}
\usage{
lohDistribution(data_frame)
}
\arguments{
\item{data_frame}{data frame with minor allele ploidy data per autosome arm, per sample}
}
\value{
None
}
\description{
Determines LOH distribution among autosomearms and cancer types an plots it in a barplot
}
\examples{
lohDistribution(results_MinorAP)
}
