% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapPlots.R
\name{lohUnsupervisedHeatmap}
\alias{lohUnsupervisedHeatmap}
\title{Plot function for unsupervised loss of heterozygosity heatmap}
\usage{
lohUnsupervisedHeatmap(results_MinorAP, mut_load_anno)
}
\arguments{
\item{results_MinorAP}{Data frame which contains minor allele ploidy data, wgd status and primary tumor location per sample}

\item{mut_load_anno}{Data frame which contains mutational load annotations per sample}
}
\value{
None
}
\description{
Plot function for unsupervised heatmap of results minor alle ploidy data
}
\examples{
lohUnsupervisedHeatmap(results_MinorAP, mut_load_anno)
}
