% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{normalization}
\alias{normalization}
\title{Ploidy normalization (genome)}
\usage{
normalization(data_frame, valNorm)
}
\arguments{
\item{data_frame}{results_CopyNumber}

\item{valNorm}{Boolean which indicates if high values need to be normalized (TRUE = yes, FALSE = no)}
}
\value{
data_frame
}
\description{
Normalizes data for the genome level so levels of aneuploidy can be detected in the heatmap plot
}
\examples{
normalization(results_CopyNumber, TRUE)
normalization(results_CopyNumber, FALSE)
}
