% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_boxPlots.R
\name{outliers}
\alias{outliers}
\title{Outlier barplot}
\usage{
outliers(data_frame, treshold)
}
\arguments{
\item{data_frame}{Can be either one of the results_ data frames}

\item{treshold}{Treshold of when a value is considered as outlier}
}
\value{
None
}
\description{
Determines for each high value (outlier) the autosome arm and cancertypes and plots it in barplot
}
\examples{
outliers(results_CopyNumber, 10)
outliers(results_MajorAP,     8)
outliers(results_MinorAP,     4)
}
