% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_proces_heatmaps.R
\name{plotCopyNumber}
\alias{plotCopyNumber}
\title{Plot copynumber data in heatmap}
\usage{
plotCopyNumber(data_frame, plotName)
}
\arguments{
\item{data_frame}{Data frame which contains copy number data, wgd status and primary tumor location per sample}

\item{plotName}{String which determines the name of the plot and if data should be split up into a smaller group based on wgdStatus}
}
\value{
None
}
\description{
Pre proccesses data to plot the copyNumber in a heatmap and calls heatMap_plot function to visualize copy number data
}
\examples{
plotCopyNumber(results_CopyNumber, plotName)
}
