% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_proces_heatmaps.R
\name{plotLOH}
\alias{plotLOH}
\title{Plot loss of heterozygosity in heatmap}
\usage{
plotLOH(results_MinorAP)
}
\arguments{
\item{results_MinorAP}{Data frame which contains minor allele ploidy data, wgd status and primary tumor location per sample}
}
\value{
None
}
\description{
Determines per cancer-type the percentage of Loss of heterozygosity for each autosome arm and calls heatmap_plot function to
visualize data in a heatmap
}
\examples{
plotLOH(results_MinorAP)
}
