% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_proces_heatmaps.R
\name{preProcesUnsupervisedHeatmap}
\alias{preProcesUnsupervisedHeatmap}
\title{Plot unsupervised heatmap of copynumber data}
\usage{
preProcesUnsupervisedHeatmap(results_CopyNumber, mut_load_anno, tp)
}
\arguments{
\item{results_CopyNumber}{Data frame which contains copy number data, wgd status and primary tumor location per sample}

\item{mut_load_anno}{Data frame which contains mutational load annotation per sample}

\item{tp}{character string which defines if data should be split}
}
\value{
None
}
\description{
Makes unsupervised heatmap of copy number data, calls plotfunction from heatmapPlots_PA.R
}
\examples{
preProcesUnsupervisedHeatmap(results_CopyNumber, mut_load_anno, "NonWGD")
preProcesUnsupervisedHeatmap(results_CopyNumber, mut_load_anno, "WGD")
preProcesUnsupervisedHeatmap(results_CopyNumber, mut_load_anno, "All")
}
