% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctions.R
\name{readAnnoCodes}
\alias{readAnnoCodes}
\title{Read annotation file with sampleID's and primary tumour locations}
\usage{
readAnnoCodes(filename)
}
\arguments{
\item{filename}{.txt file which contains sampleID column and primary tumor location column.}
}
\value{
annoCodes
}
\description{
Reads file with annotations (primaryTumorLocation of each sample) and turns it into a dataframe with sampleId and primaryTumorLocation
}
\examples{
readAnnoCodes(filename)
}
