% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFunctions.R
\name{readFile}
\alias{readFile}
\title{Read .tsv files}
\usage{
readFile(filename, annoCodes)
}
\arguments{
\item{filename}{.tsv results file from ploidyEstimator script}

\item{annoCodes}{dataframe with sampleID and primary tumor location data}
}
\value{
data_frame
}
\description{
Reads results .tsv files and makes a data frame
}
\examples{
readFile(filename, annoCodes)
}
