% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportFunctions.R
\name{wgdNormalization}
\alias{wgdNormalization}
\title{Ploidy normalization (wgd)}
\usage{
wgdNormalization(data_frame, valNorm)
}
\arguments{
\item{data_frame}{results_CopyNumber data frame}

\item{valNorm}{Boolean which indicates if high values need to be normalized (TRUE = yes, FALSE = no)}
}
\value{
data_frame
}
\description{
Normalizes data for the whole genome duplication status so levels of aneuploidy can be detected in the heatmap plot
}
\examples{
wgdNormalization(results_CopyNumber, TRUE)
wgdNormalization(results_CopyNumber, FALSE)
}
